# public/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TPLK004 - Sistem Absensi & Manajemen Akademik API",
    "description": "API Documentation untuk Sistem Absensi QR Code dengan Gamifikasi, Manajemen Tugas, dan Analytics. Sistem ini menyediakan endpoint untuk autentikasi, manajemen absensi, tugas, gamifikasi, dan berbagai fitur akademik lainnya.",
    "version": "1.0.0",
    "contact": {
      "name": "TPLK004 Support Team",
      "email": "admin@unpam.ac.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter token in format: Bearer {token}"
      },
      "sanctum": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Laravel Sanctum session authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/gamification/profile": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get gamification profile",
        "description": "Mendapatkan profil gamifikasi user (points, level, badges)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "points": {
                          "type": "integer",
                          "example": 1250
                        },
                        "level": {
                          "type": "integer",
                          "example": 5
                        },
                        "rank": {
                          "type": "integer",
                          "example": 12
                        },
                        "total_badges": {
                          "type": "integer",
                          "example": 8
                        },
                        "streak_days": {
                          "type": "integer",
                          "example": 15
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/gamification/leaderboard": {
      "get": {
        "tags": ["Gamification"],
        "summary": "Get leaderboard",
        "description": "Mendapatkan leaderboard dengan filter",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["global", "course", "class"],
              "default": "global"
            }
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["all_time", "monthly", "weekly"],
              "default": "all_time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard retrieved successfully"
          }
        }
      }
    }
  }
}